Reduce metrics cardinality (#1)
The runner_id and runner_name tag values are unique for every run
of a job. The job_workflow_ref tag value is unique for every PR
merged.  All of these result in too many metrics being returned
over time in the payload of the /metric endpoint. Eventually the
size of the payload exceeds what some agents can consume.

Likewise, this seems to contribute to constantly increasing memory
usage of the controller pod, which could ultimately lead to the pod
being oomkilled.

This change removes those high cardinality metrics.